Extract escalate_permissions, and stop shaving comments to fit
Three PRs in a row failed CI's line ceiling and I responded each time by
trimming a comment until it fit. That is treating a symptom: main.rs sat one or
two lines under its cap, so every feature tripped it.

`escalate_permissions` moves to its own module — 255 lines, and a coherent unit
on its own: it is the one place a pod's permissions widen at runtime, which is
worth being able to find. main.rs is now 4732 against a 4975 ceiling, so the next
few increments do not have to fight for space.

No behaviour change; the handler moved unmodified and the route points at it.

fmt / clippy -D warnings / cargo test --all-features (222 suites, 0 failures) /
line ratchet --strict / mediation / sealed-home / verify-strict / dep ceiling:
all green.
